rename WFController to ModelController
