+++ examples/advanced/kaplan-meier-he/src/kaplan_meier_wf_he.py
@@ -19,12 +19,12 @@ from typing import Dict
 import tenseal as ts
 
 from nvflare.app_common.abstract.fl_model import FLModel, ParamsType
-from nvflare.app_common.workflows.wf_controller import WFController
+from nvflare.app_common.workflows.model_controller import ModelController
 
 # Controller Workflow
 
 
-class KM(WFController):
+class KM(ModelController):
     def __init__(self, min_clients: int, he_context_path: str):
         super(KM, self).__init__()
         self.logger = logging.getLogger(self.__class__.__name__)
